requirements.txt
fxcmpy
python-socketio
python-dotenv
